no more opresp, controlling error with http status; bugfixes; loadbox;

+++ generators/app/templates/AppResource.js
@@ -2,6 +2,7 @@ import Vue from 'vue';
 import VueResource from 'vue-resource';
 import AppRouter from './AppRouter';
 import simpleStorage from 'simpleStorage.js';
+import AppBus from '../service/AppBus';
 
 Vue.use(VueResource);
 
@@ -13,10 +14,14 @@ Vue.http.interceptors.push((request, next) => {
     }
 
     next((resp) => {
-        if (resp.body.code === 33) {
-            simpleStorage.deleteKey("token<%= modulenameUpper %>");
-            simpleStorage.set("beforeLoginIntention", location.href);
-            AppRouter.push("/login");
+        if (!resp.ok) {
+            AppBus.$emit("alert", "danger", resp.body.message, 3000);
+            
+            if (resp.body.code === 33) {
+                simpleStorage.deleteKey("token<%= modulenameUpper %>");
+                simpleStorage.set("beforeLoginIntention", location.href);
+                AppRouter.push("/login");
+            }
         }
     });
 });
